+++ src/jsMain/kotlin/com/wadii/pages/shared/profile/EditProfileState.kt
@@ -26,9 +26,9 @@ data class EditProfileState(
     val countries: List<Country> = emptyList(),
     val provinces: List<Province> = emptyList(),
     val cities: List<City> = emptyList(),
-    val selectedCountry: Int = 0,
-    val selectedProvince: Int = 0,
-    val selectedCity: Int = 0,
+    val selectedCountry: Long = 0,
+    val selectedProvince: Long = 0,
+    val selectedCity: Long = 0,
 
     val shopName: String = "",
